+++ src/modules/cache-client/cache-client.service.ts
@@ -2,6 +2,11 @@ import axios, { AxiosError, AxiosInstance, AxiosResponse } from 'axios';
 import { stringify } from 'qs';
 import { IRoleDefinition } from '@energyweb/credential-governance';
 import { IDIDDocument } from '@ew-did-registry/did-resolver-interface';
+import {
+  Credential,
+  StatusList2021Entry,
+  VerifiableCredential,
+} from '@ew-did-registry/credentials-interface';
 import { IApp, IOrganization, IRole } from '../domains/domains.types';
 import { AssetHistory } from '../assets/assets.types';
 import {
@@ -27,6 +32,11 @@ import {
 } from './cache-client.types';
 import { SearchType } from '.';
 import { getLogger } from '../../config/logger.config';
+import {
+  RoleCredentialSubject,
+  StatusList2021Credential,
+  StatusList2021UnsignedCredential,
+} from '../verifiable-credentials';
 
 export class CacheClient implements ICacheClient {
   public pubKeyAndIdentityToken: IPubKeyAndIdentityToken | undefined;
@@ -408,6 +418,88 @@ export class CacheClient implements ICacheClient {
       `/assets/history/${id}?${query}`
     );
     return data;
+  }
+
+  /**
+   * Sets location of the credential status
+   *
+   * @param credential unsigned credential
+   * @return credential with reference to status location
+   */
+  async addStatusToCredential(
+    credential: Credential<RoleCredentialSubject>
+  ): Promise<
+    Credential<RoleCredentialSubject> & {
+      credentialStatus: StatusList2021Entry;
+    }
+  > {
+    const { data } = await this._httpClient.post<
+      Credential<RoleCredentialSubject> & {
+        credentialStatus: StatusList2021Entry;
+      }
+    >('/status-list/entries', { options: {}, credential });
+    return data;
+  }
+
+  /**
+   * Get the StatusList2021Credential object to be signed
+   *
+   * @param verifiableCredential verifiable credential to be revoked
+   * @return unsigned status list credential
+   */
+  async initiateCredentialStatusUpdate(
+    verifiableCredential: VerifiableCredential<RoleCredentialSubject>
+  ): Promise<StatusList2021UnsignedCredential> {
+    const { data } =
+      await this._httpClient.post<StatusList2021UnsignedCredential>(
+        '/status-list/credentials/status/initiate',
+        {
+          options: {},
+          verifiableCredential,
+        }
+      );
+    return data;
+  }
+
+  /**
+   * Persist signed StatusList2021Credential object in storage.
+   *
+   * @param statusListCredential signed status list
+   * @return status list credential
+   */
+  async persistCredentialStatusUpdate(
+    statusListCredential: StatusList2021Credential
+  ): Promise<StatusList2021Credential> {
+    const { data } = await this._httpClient.post<StatusList2021Credential>(
+      '/status-list/credentials/status/finalize',
+      {
+        options: {},
+        statusListCredential,
+      }
+    );
+    return data;
+  }
+
+  /**
+   * Fetch the StatusList2021Credential object from storage.
+   *
+   * @param credential verifiable credential with status list 2021
+   * @return status list credential if found
+   */
+  async getStatusListCredential(
+    credential: VerifiableCredential<RoleCredentialSubject>
+  ): Promise<StatusList2021Credential | null> {
+    if (!credential.credentialStatus?.statusListCredential) {
+      throw new Error(
+        'Missing statusListCredential property in given credential status'
+      );
+    }
+
+    const response = await axios.get<StatusList2021Credential | null>(
+      credential.credentialStatus?.statusListCredential
+    );
+
+    return response.status === 200 ? response.data : null;
   }
 
   private async refreshToken(): Promise<AuthTokens | undefined> {